requirements.txt
google-generativeai
PyPDF2
python-pptx
pandas
python-dotenv